lnwallet: move PaymentDescriptor definition to its own file
